Fixed minute instant upperbound and lowerbound bug

+++ src/time/timezone.rs
@@ -1,16 +1,16 @@
-pub const UTC_LB: i32 = -720;
-pub const UTC_UB: i32 = 840;
+pub const UTC_LB: i64 = -720;
+pub const UTC_UB: i64 = 840;
 
 /// A representation of a timezone offset, in terms of difference in minutes
 /// as compared to UTC.
 #[derive(Debug, Clone, Copy)]
-pub struct ZoneOffset(i32);
+pub struct ZoneOffset(i64);
 
 impl ZoneOffset {
   /// Creates a new ZoneOffset object according to the input offset in minutes.
   /// Only allows an offset between -12:00 and +14:00 inclusive (which is
   /// between -720 and +840 minutes).
-  pub fn new(n: i32) -> Result<Self, String> {
+  pub fn new(n: i64) -> Result<Self, String> {
     // Valid UTC offsets must be between -12:00 and +14:00, inclusive.
     if n < UTC_LB || n > UTC_UB {
       Err("Timezone offset must be a value between -720min and +840min".to_string())
@@ -24,7 +24,7 @@ impl ZoneOffset {
   }
 
   /// Returns the raw offset data.
-  pub fn raw(&self) -> i32 {
+  pub fn raw(&self) -> i64 {
     self.0
   }
 }
